fix a bug in the logic of the adding of minimum parenthesis
In order to know from which side of a binary expression or predicate
the current rewrite node was, I was comparing the current node with the
children of its parent. But the current node already contained added
parenthesis will the parent was the original (without parenthesis).
Thus the equality should be done modulo parenthesis.